General: Cancun Re Architecture
Time to rewrite Cancun initial code

+++ cancun.sh
@@ -1,33 +1,22 @@
 #!/bin/sh
 
-start() {
-    python cancun.py -s &
-    echo "Cancun Project | Started in Scheduler mode..."
-}
-
 stop() {
     pid=`ps -ef | grep 'cancun.py' | awk '{ print $2 }'`
     echo $pid
     kill $pid
     sleep 2
-    echo "Cancun Project | Scheduler mode killed!"
-    rm -f /tmp/cancun.pid
+    echo "Cancun Project Killed!"
+    test -f /tmp/cancun.pid && rm -f /tmp/cancun.pid
     /home/irlp/bin/forceunkey
 }
 
 case "$1" in
-  start)
-    start
-    ;;
   stop)
-    stop   
-    ;;
-  restart)
     stop
-    start
     ;;
   *)
-    echo "Usage: cancun.sh {start|stop|restart}"
+    echo "Usage: cancun.sh stop}"
     exit 1
 esac
+
 exit 0
